Rename files to be underscored, per Deno convention
